feat: got stuff working finally
